+++ src/modules/subscribers/subscribers.types.ts
@@ -1,0 +1,5 @@
+export interface Subscriber {
+  id: number;
+  pipeline_id: number;
+  url: string;
+}